Rename Combination Lock.java to World-CodeSprint-6/Combination Lock.java
Constraints

Each number in a slot is {0,1,2,3,4,5,6,7,8,9}.
Output Format

Print a single integer denoting the minimum number of moves to change this configuration to the correct one.

Sample Input

1 2 9 5 7
1 3 2 0 7
Sample Output

9
Explanation

We perform the following operations on each slot:

Slot : We rotate the dial 0 times because this slot is already displaying the desired number (i.e., ).
Slot : We rotate the dial 1 time, changing it from 2->3.
Slot : We rotate the dial 3 times, changing it from 9->0->1->2 .
Slot : We rotate the dial 5 times, changing it from 5->4->3->2->1->0.
Slot : We rotate the dial 0 times because this slot is already displaying the desired number (i.e., ).
Finally, we sum the number of operations performed at each slot: 0+1+3+5+0=9. Thus, we print 9 as our answer.