Rename Bon Appétit.java to World-CodeSprint-6/Bon Appétit.java
Sample Input 0

4 1
3 10 2 9
12
Sample Output 0

5
Explanation 0
Anna didn't eat item c[1]=10, but she shared the rest of the items 3+2+9=14 with Brian. The total cost of the shared items is  and, split in half, the cost per person is b[actual] =7  . Brian charged her b[charged] = 12 for her portion of the bill, which is more than the 7 dollars worth of food that she actually shared with him. Thus, we print the amount Anna was overcharged, b[charged] - b[actual] = 12 - 7 = 5, on a new line.

Sample Input 1

4 1
3 10 2 9
7
Sample Output 1

Bon Appetit
Explanation 1
Anna didn't eat item c[1]=10, but she shared the rest of the items 3+2+9=14with Brian. The total cost of the shared items is  and, split in half, the cost per person is b[actual] =7. Because this matches the amount, b[charged] = 7, that Brian charged Anna for her portion of the bill, we print Bon Appetit on a new line.